% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_sim_ind.R
\name{attribute_sim_ind}
\alias{attribute_sim_ind}
\title{Simulate the attributes for each shape and each level with the assumption
that the attributes of shapes are independent of the attributes of levels}
\usage{
attribute_sim_ind(
  product_shapes_and_levels,
  attributes_number,
  shape_attributes_number,
  level_attributes_number
)
}
\arguments{
\item{product_shapes_and_levels}{A numeric dateframe of three colunms: product_id, assigned_shape and assigned_level}

\item{attributes_number}{The number of attributes}

\item{shape_attributes_number}{The number of attributes assigned to shape}

\item{level_attributes_number}{The number of attributes assigned to level}
}
\value{
A numeric dateframe of three colunms: product_id, assigned_shape and assigned_level and attributes
}
\description{
Simulate the attributes for each shape and each level with the assumption
that the attributes of shapes are independent of the attributes of levels
}
\examples{
attribute_sim_ind(product_shapes_and_levels=data.frame(product_id=1:3,assigned_shape=c(1,1,2), assigned_level=c(5,3,1)),
attributes_number=15,
shape_attributes_number=7,
level_attributes_number=4)

}
