% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdsim_bass.R
\name{npdsim_bass}
\alias{npdsim_bass}
\title{Calculate the Bass probability}
\usage{
npdsim_bass(p_param, q_param, t)
}
\arguments{
\item{p_param}{Coefficient of innovation}

\item{q_param}{Coefficient of imitation}

\item{t}{A numeric vector of time periods}
}
\value{
A numeric vector of the probability density function of purchase at time t, f(t)
}
\description{
Calculate the Bass probability density function of purchase f(t)
}
\examples{
npdsim_bass(p_param=0.01,q_param=0.2, t=1:20)
}
