% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_sim_ig.R
\name{piecewise_shape_sim}
\alias{piecewise_shape_sim}
\title{Generate piece-wise linear (4 segments) generic shapes for the introduction and growth phases}
\usage{
piecewise_shape_sim(periods_number, shape_number)
}
\arguments{
\item{periods_number}{Number of periods of the introduction and growth phases}

\item{shape_number}{Number of generic shapes}
}
\value{
A numeric dateframe of three colunms: time, shape and assigned_shape
}
\description{
Generate piece-wise linear (4 segments) generic shapes for the introduction and growth phases
}
\examples{
piecewise_shape_sim(periods_number=20, shape_number=5)

}
