% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_sim_bass.R
\name{shape_sim_bass}
\alias{shape_sim_bass}
\title{Generate generic Bass shapes}
\usage{
shape_sim_bass(periods_number, shape_number)
}
\arguments{
\item{periods_number}{Number of time periods of the products life cycle}

\item{shape_number}{Number of generic shapes}
}
\value{
A numeric dateframe of three columns: time, shape and assigned_shape
}
\description{
Generate generic Bass shapes for the demand of new products during their life cycle
}
\examples{
shape_sim_bass(periods_number=20, shape_number=5)
}
