% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_sim_random.R
\name{shape_sim_random}
\alias{shape_sim_random}
\title{Generate random (Bass, Trapezoidal or Triangular) shapes for the demand of new products during their life cycle}
\usage{
shape_sim_random(periods_number, shape_number)
}
\arguments{
\item{periods_number}{Number of time periods of the products life cycle}

\item{shape_number}{Number of generic shapes}
}
\value{
A numeric dateframe of three colunms: time, shape and assigned_shape
}
\description{
Generate random (Bass, Trapezoidal or Triangular) shapes for the demand of new products during their life cycle
}
\examples{
shape_sim_random(periods_number=20, shape_number=5)
}
