% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_sim_triangle.R
\name{shape_sim_triangle}
\alias{shape_sim_triangle}
\title{Generate triangular shapes}
\usage{
shape_sim_triangle(periods_number, shape_number)
}
\arguments{
\item{periods_number}{Number of time periods of the products life cycle}

\item{shape_number}{Number of generic shapes}
}
\value{
A numeric dateframe of three colunms: time, shape and assigned_shape
}
\description{
Generate triangular shapes for the demand of new products
during their life cycle
}
\examples{
shape_sim_triangle(periods_number=20, shape_number=5)
}
